refactor(timeInterval): moved out of extended directory
BREAKING CHANGE: `timeInterval` has been moved out of `extended` folder and under `operator` folder directly
