+++ src/Rx.KitchenSink.ts
@@ -120,7 +120,7 @@ import './add/operator/takeUntil';
 import './add/operator/takeWhile';
 import './add/operator/throttle';
 import './add/operator/throttleTime';
-import './add/operator/extended/timeInterval';
+import './add/operator/timeInterval';
 import './add/operator/timeout';
 import './add/operator/timeoutWith';
 import './add/operator/toArray';
@@ -149,7 +149,7 @@ import {asap} from './scheduler/asap';
 import {queue} from './scheduler/queue';
 import {AsapScheduler} from './scheduler/AsapScheduler';
 import {QueueScheduler} from './scheduler/QueueScheduler';
-import {TimeInterval} from './operator/extended/timeInterval';
+import {TimeInterval} from './operator/timeInterval';
 import {TestScheduler} from './testing/TestScheduler';
 import {VirtualTimeScheduler} from './scheduler/VirtualTimeScheduler';
 import {rxSubscriber} from './symbol/rxSubscriber';
