+++ src/operator/timeInterval.ts
@@ -1,8 +1,8 @@
-import {Operator} from '../../Operator';
-import {Observable} from '../../Observable';
-import {Subscriber} from '../../Subscriber';
-import {Scheduler} from '../../Scheduler';
-import {queue} from '../../scheduler/queue';
+import {Operator} from '../Operator';
+import {Observable} from '../Observable';
+import {Subscriber} from '../Subscriber';
+import {Scheduler} from '../Scheduler';
+import {queue} from '../scheduler/queue';
 
 export function timeInterval<T>(scheduler: Scheduler = queue): Observable<TimeInterval> {
   return this.lift(new TimeIntervalOperator(scheduler));
